Endpoint de Estatisticas já está com a regra de 60 segundos e utilizando a função DoubleSummaryStatistics

+++ src/main/java/com/jan1ooo/apirest/model/statistics/Estatisticas.java
@@ -5,17 +5,15 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.math.BigDecimal;
-
 @Getter
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
 public class Estatisticas {
 
-    private Integer count;
-    private BigDecimal sum;
-    private BigDecimal avg;
+    private Long count;
+    private Double sum;
+    private Double avg;
     private Double min;
     private Double max;
 }
